+++ src/repo/lock.rs
@@ -2,11 +2,22 @@
 //!
 //! Consists of [`FsDataStore`] and [`FsBlockStore`].
 
+#[cfg(not(target_arch = "wasm32"))]
+use std::sync::Arc;
+
+use either::Either;
+
 use super::{Lock, LockError};
 
 #[cfg(not(target_arch = "wasm32"))]
-#[derive(Debug)]
+#[derive(Clone, Debug)]
 pub struct FsLock {
+    inner: Arc<InnerFsLock>,
+}
+
+#[cfg(not(target_arch = "wasm32"))]
+#[derive(Debug)]
+struct InnerFsLock {
     file: parking_lot::Mutex<Option<std::fs::File>>,
     path: std::path::PathBuf,
     state: parking_lot::Mutex<State>,
@@ -23,9 +34,11 @@ enum State {
 impl FsLock {
     pub fn new(path: std::path::PathBuf) -> Self {
         Self {
-            file: parking_lot::Mutex::new(None),
-            path,
-            state: parking_lot::Mutex::new(State::Unlocked),
+            inner: Arc::new(InnerFsLock {
+                file: parking_lot::Mutex::new(None),
+                path,
+                state: parking_lot::Mutex::new(State::Unlocked),
+            }),
         }
     }
 }
@@ -41,23 +54,32 @@ impl Lock for FsLock {
             .write(true)
             .create(true)
             .truncate(true)
-            .open(&self.path)?;
+            .open(&self.inner.path)?;
 
         file.try_lock_exclusive()?;
 
-        *self.state.lock() = State::Exclusive;
-        *self.file.lock() = Some(file);
+        *self.inner.state.lock() = State::Exclusive;
+        *self.inner.file.lock() = Some(file);
 
         Ok(())
     }
 }
 
-#[derive(Debug, Default)]
+#[derive(Clone, Copy, Debug, Default)]
 pub struct MemLock;
 
 impl Lock for MemLock {
     fn try_exclusive(&self) -> Result<(), LockError> {
         Ok(())
+    }
+}
+
+impl<L: Lock, R: Lock> Lock for Either<L, R> {
+    fn try_exclusive(&self) -> Result<(), LockError> {
+        match self {
+            Either::Left(lock) => lock.try_exclusive(),
+            Either::Right(lock) => lock.try_exclusive(),
+        }
     }
 }